a program that finds and prints the first 98 Fibonacci numbers, starting with 1 and 2

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,14 +10,14 @@
 int main(void)
 {
 	int i;
-	unsigned long long f1 = 1, f2 = 2, n;
+	unsigned long f1 = 1, f2 = 2, n;
 
-	printf("%llu, %llu", f1, f2);
+	printf("%lu, %lu", f1, f2);
 
 	for (i = 3; i <= 98; i++)
 	{
 		n = f1 + f2;
-		printf(", %llu", n);
+		printf(", %lu", n);
 		f1 = f2;
 		f2 = n;
 	}
